refactor: remove unused recoil states
