+++ frontend/src/components/Teams/CreateTeam/CreateTeam.tsx
@@ -17,7 +17,8 @@ import TeamMembersList from '@/components/Teams/Team/TeamMembersList';
 import useCreateTeam from '@/hooks/teams/useCreateTeam';
 import useCurrentSession from '@/hooks/useCurrentSession';
 import SchemaCreateTeam from '@/schema/schemaCreateTeamForm';
-import { createTeamState, usersListState } from '@/store/team/atom/team.atom';
+import { createTeamState } from '@/store/team.atom';
+import { usersListState } from '@/store/user.atom';
 import { StyledForm } from '@/styles/pages/pages.styles';
 import { CreateTeamUser } from '@/types/team/team.user';
 import { ROUTES } from '@/utils/routes';
