+++ frontend/src/hooks/useBoardUtils.tsx
@@ -3,8 +3,8 @@ import { useSession } from 'next-auth/react';
 import { useQueryClient } from '@tanstack/react-query';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 
-import { usersListState } from '@/store/team/atom/team.atom';
 import { toastState } from '@/store/toast/atom/toast.atom';
+import { usersListState } from '@/store/user.atom';
 
 const useBoardUtils = () => {
   const router = useRouter();
