+++ frontend/src/pages/boards/newRegularBoard.tsx
@@ -25,8 +25,8 @@ import useBoard from '@/hooks/useBoard';
 import useCurrentSession from '@/hooks/useCurrentSession';
 import SchemaCreateRegularBoard from '@/schema/schemaCreateRegularBoard';
 import { createBoardDataState, createBoardTeam } from '@/store/createBoard/atoms/create-board.atom';
-import { usersListState } from '@/store/team/atom/team.atom';
 import { toastState } from '@/store/toast/atom/toast.atom';
+import { usersListState } from '@/store/user.atom';
 import { StyledForm } from '@/styles/pages/pages.styles';
 import { BoardUserDto } from '@/types/board/board.user';
 import { TEAMS_KEY } from '@/utils/constants/reactQueryKeys';
