+++ frontend/src/pages/teams/new.tsx
@@ -13,7 +13,8 @@ import LoadingPage from '@/components/Primitives/Loading/Page/Page';
 import CreateTeam from '@/components/Teams/CreateTeam/CreateTeam';
 import useCurrentSession from '@/hooks/useCurrentSession';
 import useUsers from '@/hooks/users/useUsers';
-import { createTeamState, usersListState } from '@/store/team/atom/team.atom';
+import { createTeamState } from '@/store/team.atom';
+import { usersListState } from '@/store/user.atom';
 import { USERS_KEY } from '@/utils/constants/reactQueryKeys';
 import { TeamUserRoles } from '@/utils/enums/team.user.roles';
 import { ROUTES } from '@/utils/routes';
