+++ frontend/src/store/user.atom.tsx
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+import { UserList } from '@/types/team/userList';
+
+export const usersListState = atom<UserList[]>({
+  key: 'usersList',
+  default: [],
+});
