--- frontend/src/store/user/selectors/get-user.selector.tsx
@@ -1,8 +1,0 @@
-import { selector } from 'recoil';
-
-import { userState } from '@/store/user/atoms/user.atom';
-
-export const getUserSelector = selector({
-  key: 'getUser', // unique ID (with respect to other atoms/selectors)
-  get: ({ get }) => get(userState),
-});
