--- frontend/src/store/vote/atoms/vote.atom.tsx
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const maxVotesReachedAtom = atom<{ maxVotesReached: boolean; noVotes: boolean }>({
-  key: 'maxVotesReachedAtom',
-  default: { maxVotesReached: false, noVotes: false },
-});
